shop_taxes/requirements.txt
django==1.4.3
django-bootstrap3
django-couch-utils
pycurl
django-annoying
couchdbcurl